Factor out main.
Because we use tsne through the cython wrapper, we don't actually use
the c++ main method.
