+++ tsne/bh_sne_src/tsne.cpp
@@ -727,51 +727,3 @@ void TSNE::save_csv(const char* csv_file, double* Y, int N, int D) {
 
     csv.close();
 }
-
-// Function that runs the Barnes-Hut implementation of t-SNE
-int main(int argc, char *argv[]) {
-
-    // load input and output
-    std::string dat_file = "data.dat";
-    std::string res_file = "result.dat";
-    if (argc > 1) {
-        dat_file = argv[1];
-        res_file = argv[2];
-    }
-
-    const char *dat_file_c = dat_file.c_str();
-    const char *res_file_c = res_file.c_str();
-
-    // Define some variables
-	int origN, N, D, no_dims, max_iter, *landmarks;
-	double perc_landmarks;
-	double perplexity, theta, *data;
-    int rand_seed = -1;
-    TSNE* tsne = new TSNE();
-
-    // Read the parameters and the dataset
-	if(tsne->load_data(dat_file_c, &data, &origN, &D, &no_dims, &theta, &perplexity, &rand_seed, &max_iter)) {
-
-		// Make dummy landmarks
-        N = origN;
-        int* landmarks = (int*) malloc(N * sizeof(int));
-        if(landmarks == NULL) { fprintf(stderr,"Memory allocation failed!\n"); exit(1); }
-        for(int n = 0; n < N; n++) landmarks[n] = n;
-
-		// Now fire up the SNE implementation
-		double* Y = (double*) malloc(N * no_dims * sizeof(double));
-		double* costs = (double*) calloc(N, sizeof(double));
-        if(Y == NULL || costs == NULL) { fprintf(stderr,"Memory allocation failed!\n"); exit(1); }
-	    tsne->run(data, N, D, Y, no_dims, perplexity, theta, rand_seed, false, NULL, false, max_iter);
-
-		// Save the results
-		tsne->save_data(res_file_c, Y, landmarks, costs, N, no_dims);
-
-        // Clean up the memory
-		free(data); data = NULL;
-		free(Y); Y = NULL;
-		free(costs); costs = NULL;
-		free(landmarks); landmarks = NULL;
-    }
-    delete(tsne);
-}
